+++ src/components/Friendly-captcha/Frontend/useCaptchaHook.tsx
@@ -136,23 +136,21 @@ const FriendlyCaptcha = (props: FriendlyCaptchaWidgetProps) => {
       {!props.showAttribution && (
         <style>{'.frc-banner { display: none }'}</style>
       )}
-      {props.customWidgetStyle && (
-        <style>
-          {props.customWidgetStyle.icon &&
-            `#use-friendly-captcha-container .frc-icon {${cssToString(
-              props.customWidgetStyle.icon
-            )}}`}
-          {props.customWidgetStyle.button &&
-            `#use-friendly-captcha-container .frc-button {${cssToString(
-              props.customWidgetStyle.button
-            )}}`}
-          {props.customWidgetStyle.text &&
-            `#use-friendly-captcha-container .frc-text {${cssToString(
-              props.customWidgetStyle.text
-            )}}`}
-        </style>
+      {props.customWidgetStyle?.icon && (
+        <style>{`#use-friendly-captcha-container .frc-icon {${cssToString(
+          props.customWidgetStyle.icon
+        )}}`}</style>
       )}
-
+      {props.customWidgetStyle?.button && (
+        <style>{`#use-friendly-captcha-container .frc-button {${cssToString(
+          props.customWidgetStyle.button
+        )}}`}</style>
+      )}
+      {props.customWidgetStyle?.text && (
+        <style>{`#use-friendly-captcha-container .frc-text {${cssToString(
+          props.customWidgetStyle.text
+        )}}`}</style>
+      )}
       <div
         ref={container}
         id="use-friendly-captcha-container"