+++ CHANGELOG.md
@@ -22,3 +22,9 @@
 
 ### v1.2.1
 - Update range of possible css attributes to style the captcha widget
+
+### v1.2.2
+- Fix nextjs hydration error, that occurs when rendering the css for the captcha widget
+
+### v1.2.3
+- Adjust nextjs hydration for style tags